Score system: plumb BigInt end-to-end
The fractal world scales forever — a 20-min sim already reaches ~5.8e20
points, way past Number.MAX_SAFE_INTEGER (9e15), so Number scoring silently
loses integer precision. Convert the authoritative score/points path to
BigInt while keeping physics (radius, level, speed) as Number.

- catalog.pointsFor returns BigInt
- hole.score starts 0n; eat() does BigInt += BigInt
- swallow.finalizeSwallow: obj.points * BigInt(sw.mult)
- hud.fmtNum accepts Number or BigInt, walks a suffix ladder
  ['', K, M, B, T, Qa, Qi, Sx, Sp, Oc, No, Dc], and past Dc (1e33) falls
  back to mantissa+exponent so any BigInt magnitude renders
- hud.saveBest/loadBest: serialize score as decimal string; comparison
  via BigInt() handles legacy Number-form bests too
- hud.update easing: keep displayScore as a Number approximation (ease
  toward Number(hole.score); snap past 1e15 where float steps get huge)
- e2e smoke: cross the page.evaluate boundary as strings; write 4321n
  inside evaluate; expect stored score "4321"

tools/simulate.mjs needs no code change: BigInt in template literals and
Array#join renders as a fixed-notation decimal, so the CSV column that
used to show "5.86e+20" now shows exact integer digits.

+++ js/catalog.js
@@ -167,8 +167,12 @@ export function sizeMultForBand(band) {
 }
 
 // Points scale with placed (post-multiplier) area — scale-free across cycles.
+// Returns BigInt: the fractal world can produce scores well past 2^53, so the
+// authoritative point value has to be exact-integer at any magnitude. The
+// double-precision area math before conversion is fine — magnitude is what
+// matters, not the last few digits of an already-huge number.
 export function pointsFor(placedRadius) {
-  return Math.max(1, Math.round((placedRadius * placedRadius) / CONFIG.POINTS_DIV));
+  return BigInt(Math.max(1, Math.round((placedRadius * placedRadius) / CONFIG.POINTS_DIV)));
 }
 
 // Roman-numeral suffix for cycled biome names: "Berry Meadow II".
